+++ esp-hal-embassy/src/fmt.rs
@@ -1,9 +1,6 @@
 #![macro_use]
 #![allow(unused_macros)]
 
-#[cfg(all(feature = "defmt", feature = "log"))]
-compile_error!("You may not enable both `defmt` and `log` features.");
-
 #[collapse_debuginfo(yes)]
 macro_rules! assert {
     ($($x:tt)*) => {